Demo/requirements.txt
streamlit>=1.36.0
plotly>=5.24.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
